Add files via upload
Simplest overview, in each process:

Read in the data from the input files, put into unverified blocks (UBs).
Multicast (send) the UBs to all the processes (in JSON format).
Do the work of verifying a UB. Add it to the blockchain (BC), multicast the new BC to all the processes (in JSON format).
Repeat until done. Write the BC to disk (in JSON format).